Naming, location, and documentation updates
