+++ demo/test/snapshots/linting-config.test.js.md
@@ -1,10 +1,10 @@
-# Snapshot report for `demo/test/lint-output.js`
+# Snapshot report for `demo/test/linting-config.test.js`
 
-The actual snapshot is saved in `lint-output.js.snap`.
+The actual snapshot is saved in `linting-config.test.js.snap`.
 
 Generated by [AVA](https://ava.li).
 
-## Should have consistent eslint final config
+## Should apply a consistent overall eslint configuration
 
 > Snapshot 1
 
@@ -1295,7 +1295,7 @@ Generated by [AVA](https://ava.li).
       "ignorePatterns": []␊
     }`
 
-## Should have consistent rule output
+## Should apply our custom linting rules consistently
 
 > Snapshot 1
 
@@ -1365,9 +1365,6 @@ Generated by [AVA](https://ava.li).
       2:3  warning  Unexpected var, use let or const instead  no-var␊
       3:3  warning  Unexpected var, use let or const instead  no-var␊
     ␊
-    test/lint-output.js␊
-      12:7  warning  Unexpected console statement  no-console␊
-    ␊
-    ✖ 60 problems (25 errors, 35 warnings)␊
+    ✖ 59 problems (25 errors, 34 warnings)␊
       0 errors and 2 warnings potentially fixable with the `--fix` option.␊
     `
